requirements.txt
langchain
langchain_community
langchain_huggingface
python-dotenv
